wip(transl) add SNI match criteria to routes
Add SNI match criteria to Kong routes when the source Ingress has a TLS
rule with hostnames.

This change always adds this configuration if possible from the Ingress
object. This introduces two possible issues:

- If any Kong functionality differs for routes with no SNI information
  but with hostname information, this change will break that
  functionality for controller-managed routes. This likely applies to
  any controller-managed routes that expect clients with older TLS
  stacks, i.e. clients whose TLS stacks lack SNI support but will still
  send well-formed HTTP/1.1 requests including a Host header.
- This does not allow any Kong functionality that relies on the ability
  to specify an SNI match criterion that does not correspond to some
  hostname match criterion, possibly introducing murky scenarios where
  wildcard route hostnames are not the same as wildcard route SNIs.
  Expected impact here is limited, but not out of the question.

As-is, practical testing indicates this works as expected. This remains
a WIP because it lacks unit tests and does not attempt to account for
the edge cases above.